地面に対する、ポイントライトの見え方がおかしくなってしまった（法線が怪しい）。 武器のUIを作った。 キー入力を一つのコンフィグに対し、複数のキーを割り当てることができるようにした。

+++ DX11MyEngine/Source/Cpp/Component_Light.cpp
@@ -1,8 +1,8 @@
 #include "pch.h"
 #include "Component_Light.h"
 
-constexpr int DEF_INTENSITY = 1.0f;
-constexpr int DEF_COLOR[3] = { 1.0f,1.0f,1.0f };
+constexpr float DEF_INTENSITY = 1.0f;
+constexpr float DEF_COLOR[3] = { 1.0f,1.0f,1.0f };
 
 //*---------------------------------------------------------------------------------------
 //* @:Light Class
